july14/requirements.txt
quart
load_dotenv
azure.cognitiveservices.speech
httpx
asyncio
time
openai
threading
queue
AzureOpenAI
azure-cognitiveservices-speech
asyncio
concurrent.futures
logging
concurrent.future
quart
aiohttp